Changes to CSS and deleted some social links.
Changed the size of the profile picture and adjusted the border.
Played around with the color scheme of the page.
Changed the background color of the cards and adjusted the focus color.
Changed the <li> color scheme on the goals card.
Added new instagram and github icons to the footer.